Move backoffice regulation files to directory
