+++ frontend/src/App.js
@@ -17,7 +17,7 @@ import UpdatingVesselLoader from './features/vessel_sidebar/UpdatingVesselLoader
 import RightMenuOnHoverArea from './features/vessel_sidebar/RightMenuOnHoverArea'
 import Measurement from './features/measurements/Measurement'
 import VesselFilters from './features/vessel_filters/VesselFilters'
-import EditRegulation from './features/backoffice/create_regulation/EditRegulation'
+import EditRegulation from './features/backoffice/edit_regulation/EditRegulation'
 import { ReactComponent as AlertSVG } from './features/icons/Picto_alerte.svg'
 import { Provider } from 'react-redux'
 import { backofficeStore, homeStore, backofficePersistor } from './Store'
