+++ frontend/src/features/backoffice/Backoffice.js
@@ -3,8 +3,8 @@ import { batch, useDispatch, useSelector } from 'react-redux'
 import styled from 'styled-components'
 
 import BaseMap from '../map/BaseMap'
-import LawType from './LawType'
-import SearchRegulations from './SearchRegulations'
+import LawType from './list_regulation/LawType'
+import SearchRegulations from './list_regulation/SearchRegulations'
 import RegulatoryZoneMetadata from '../layers/regulatory/RegulatoryZoneMetadata'
 import BaseLayer from '../../layers/BaseLayer'
 import RegulatoryLayers from '../../layers/RegulatoryLayers'
