+++ frontend/src/features/backoffice/edit_regulation/AuthorizedRadioButtonGroup.js
