+++ frontend/src/features/backoffice/edit_regulation/ConfirmRegulationModal.js
