+++ frontend/src/features/backoffice/edit_regulation/EditRegulation.js
