+++ frontend/src/features/backoffice/edit_regulation/InfoBox.js
