+++ frontend/src/features/backoffice/edit_regulation/InfoPoint.js
