+++ frontend/src/features/backoffice/edit_regulation/custom_form/CustomSelectComponent.js
