+++ frontend/src/features/backoffice/edit_regulation/fishing_period/HolidayCheckbox.js
