+++ frontend/src/features/backoffice/edit_regulation/gear_regulation/GearRegulation.js
