+++ frontend/src/features/backoffice/edit_regulation/gear_regulation/RegulatedGears.js
