+++ frontend/src/features/backoffice/edit_regulation/identification/RegulationLawTypeLine.js
