+++ frontend/src/features/backoffice/edit_regulation/index.js
