+++ frontend/src/features/backoffice/edit_regulation/regulatory_species/SpeciesRegulation.js
