+++ frontend/src/features/backoffice/edit_regulation/regulatory_text/RegulatoryTextSection.js
